% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaReg.R
\name{AdaReg}
\alias{AdaReg}
\title{AdaReg}
\usage{
AdaReg(
  X,
  y,
  gam.seq = seq(0, 3, by = 0.1),
  var.gp.id = NULL,
  mu.fix.value = NULL,
  var.fix.value = NULL,
  tol = 10^(-4),
  step = 50
)
}
\arguments{
\item{X}{The design matrix in linear regression.}

\item{y}{The response variable in linear regression.}

\item{gam.seq}{A sequence of gamma's (default: seq(0, 3, by=.1)).}

\item{var.gp.id}{A vector of groups indicating the samples in the same variance group (default: all the samples have the same variance).}

\item{tol}{Tolerance for interations (default: 10^(-4)).}

\item{step}{Step limit (default: 50).}
}
\value{
Estimated coef and variance, and fitting info for the residuals.
}
\description{
To robustly estimate linear regression coefficients and detect outliers.
}
\author{
Meng Wang
\email{mengw1@stanford.edu}
}
