% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{AIPW}
\alias{AIPW}
\title{AIPW estimator}
\usage{
AIPW(
  data,
  hyperparams,
  reg_type = "lasso",
  return_reg = TRUE,
  first_batch = FALSE
)
}
\arguments{
\item{data}{Data containing the reward, arm and covariate}

\item{hyperparams}{The hyperparameters associated to the regression, an input to fit_glm and fit_lasso}

\item{reg_type}{Regression methods, e.g., lasso}

\item{return_reg}{Return the regression coefficient or not}

\item{first_batch}{TRUE if only data in first batch is sued; FALSE if full data is sued}
}
\value{
A scalar used for downstream inference
}
\description{
AIPW estimator
}
