% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BDM_test}
\alias{BDM_test}
\title{Batchwise Difference-in-Means (BDM) test}
\usage{
BDM_test(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided p-values
}
\description{
Normalizes the difference-in-means per batch and aggregates across batches.
Under the null, the test statistic is asymptotically standard normal.
See Hirano and Porter (2025), Appendix D.
}
