% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{CI_weighted_AIPW}
\alias{CI_weighted_AIPW}
\title{A function providing confidence interval for weighted_AIPW}
\usage{
CI_weighted_AIPW(
  data,
  hyperparams,
  reg_type,
  estimand = "ATE",
  arm,
  sig_level = 0.05
)
}
\arguments{
\item{data}{Data as input same as in CI_AIPW function}

\item{hyperparams}{Same as CI_AIPW function}

\item{reg_type}{Same as CI_AIPW function}

\item{estimand}{Same as CI_AIPW function}

\item{arm}{Same as CI_AIPW function}

\item{sig_level}{Same as CI_AIPW function}
}
\value{
Same as CI_AIPW function
}
\description{
A function providing confidence interval for weighted_AIPW
}
