% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{covariance_estimate}
\alias{covariance_estimate}
\title{Covariance estimate}
\usage{
covariance_estimate(conditional_mean, mean_estimate)
}
\arguments{
\item{conditional_mean}{Conditional mean vector of length n}

\item{mean_estimate}{Vector of mean estimate of two arms}
}
\value{
Covariance estimate
}
\description{
Covariance estimate
}
