% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{data_generate_online}
\alias{data_generate_online}
\title{Data generation function}
\usage{
data_generate_online(
  eps,
  dist_fam,
  n_1,
  n_2,
  p,
  coef_list = NULL,
  deciding_scheme = "IPW",
  reg_type = NULL,
  hyperparams = NULL,
  dist_params,
  initial_prob,
  link = "linear",
  type = "eps_greedy"
)
}
\arguments{
\item{eps}{The input for sampling algorithm}

\item{dist_fam}{A vector storing the type of sampling distribution}

\item{n_1}{Number of sample in stage 1}

\item{n_2}{Number of sample in stage 2}

\item{p}{Dimension of the covariate}

\item{coef_list}{List of coefficients}

\item{deciding_scheme}{IPW or AIPW}

\item{reg_type}{NULL as default choice}

\item{hyperparams}{Null as default choice}

\item{dist_params}{A matrix storing the hyperparameters for sampling}

\item{initial_prob}{A vector storing the initial sampling probability for each arm}

\item{link}{When it is linear, linear model is considered; when it is nonlinear, nonlinear model is considered}

\item{type}{sampling algorithm choice; either epsilon greedy or Thompson sampling}
}
\value{
Data list containing reward, arm, covariate and batch ids
}
\description{
Data generation function
}
