% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{fit_glm}
\alias{fit_glm}
\title{Function fitting GLM regression}
\usage{
fit_glm(data, hyperparams)
}
\arguments{
\item{data}{Data containing covariate and reward for the specific arm}

\item{hyperparams}{Hyperparmeters containing the family}
}
\value{
Same output as in fit_lasso function
}
\description{
Function fitting GLM regression
}
