% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{fit_lasso}
\alias{fit_lasso}
\title{lasso fit}
\usage{
fit_lasso(data, hyperparams)
}
\arguments{
\item{data}{Data containing covariate and reward}

\item{hyperparams}{Hyperparameter for lasso regression}
}
\value{
A list of output containing the estimated coef, conditional mean etc
}
\description{
lasso fit
}
