% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{IPW}
\alias{IPW}
\title{This is a function for IPW point estimate}
\usage{
IPW(data, one_batch = FALSE, batch_output = 1)
}
\arguments{
\item{data}{Data containing reward, arm, batch_id and sampling probability}

\item{one_batch}{A logical variable deciding if only one of the batches is only used}

\item{batch_output}{Which batch result will be output}
}
\value{
A vector of estimated mean and variance for each arm
}
\description{
This is a function for IPW point estimate
}
