% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{limiting_sampling_function}
\alias{limiting_sampling_function}
\title{Generating random variable in the limiting distribution}
\usage{
limiting_sampling_function(
  A_1,
  V_1,
  h,
  H_1,
  eps,
  type = "eps_greedy",
  weight_vec = c(1, -1)
)
}
\arguments{
\item{A_1}{First stage bivariate limiting normal distribution}

\item{V_1}{Variance estimate in the first stage}

\item{h}{Signal (c)}

\item{H_1}{Limiting weight vector in the first stage}

\item{eps}{Epsilon parameter in epsilon-greedy algorithm}

\item{type}{either epsilon greedy algorithm or Thompson sampling}

\item{weight_vec}{A weight vector depending how ATE is computed in sampling function (encourage or discourage)}
}
\value{
Probability of choosing specific arm
}
\description{
Generating random variable in the limiting distribution
}
