% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NA_test}
\alias{NA_test}
\title{Normalized adaptive weighting test}
\usage{
NA_test(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided tests
}
\description{
Normalized adaptive weighting test
}
