% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NC_test}
\alias{NC_test}
\title{Normalized constant weighting test}
\usage{
NC_test(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided tests
}
\description{
Normalized constant weighting test
}
