% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NDM_test}
\alias{NDM_test}
\title{Test with normalized difference-in-means test statistic}
\usage{
NDM_test(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided tests
}
\description{
Test with normalized difference-in-means test statistic
}
