% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{normal_test}
\alias{normal_test}
\title{IPW test with normal quantile}
\usage{
normal_test(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided tests
}
\description{
IPW test with normal quantile
}
