% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{plugin_bootstrap}
\alias{plugin_bootstrap}
\title{Bootstrap function for constant weighting unnormalized test statistic}
\usage{
plugin_bootstrap(
  mean_estimate,
  variance_estimate,
  eps,
  fs_p,
  B = 2000,
  normalization = "normalized",
  h = 0,
  weighting = "aw",
  type = "eps_greedy",
  weight_vec = c(1, -1)
)
}
\arguments{
\item{mean_estimate}{Estimate for the expected outcome (bivariate vector)}

\item{variance_estimate}{Variance estimate (bivariate vector)}

\item{eps}{Epsilon tolerance in epsilon-greedy algorithm}

\item{fs_p}{Treatment assignment probability in the first stage for the first arm}

\item{B}{Number of bootstrap sample required}

\item{normalization}{A hyperparameter to use normalization or not in the test statistic}

\item{h}{Signal strength present in the sampling function}

\item{weighting}{A hyperparameter to use adaptive or constant weighting}

\item{type}{Sampling algorithm}

\item{weight_vec}{A weight vector depending how ATE is computed in sampling function (encourage or discourage)}
}
\value{
A vector of bootstrap sample of length B
}
\description{
Bootstrap function for constant weighting unnormalized test statistic
}
