% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{sampling_function}
\alias{sampling_function}
\title{Sampling algorithm}
\usage{
sampling_function(
  eps,
  reward,
  n,
  dist_fam = NULL,
  dist_params = NULL,
  data_generation = TRUE,
  sample_size = NULL,
  type = "eps_greedy"
)
}
\arguments{
\item{eps}{A small probability that is used to draw random sample from arms}

\item{reward}{A vector of aggregated rewards to determine which arm to sample}

\item{n}{Number of data that is needed to sample}

\item{dist_fam}{Distribution that is used for sampling}

\item{dist_params}{Parameters for the distribution, e.g., mean or variance}

\item{data_generation}{A logical value determing if we want to generate the data or not}

\item{sample_size}{sample size used to calculate expected reward estimate; can equal n}

\item{type}{sampling algorithm}
}
\value{
the random sampled and random index with length n
}
\description{
Sampling algorithm
}
