% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{second_batch_only}
\alias{second_batch_only}
\title{Test with sample splitting}
\usage{
second_batch_only(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided p-values
}
\description{
Test with sample splitting
}
