% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{UA_test}
\alias{UA_test}
\title{Unnormalized adaptive weighting test}
\usage{
UA_test(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided tests
}
\description{
Unnormalized adaptive weighting test
}
