% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{UC_test}
\alias{UC_test}
\title{Unnormalized constant weighting test}
\usage{
UC_test(data)
}
\arguments{
\item{data}{Data to be analyed}
}
\value{
left, right and both sided tests
}
\description{
Unnormalized constant weighting test
}
