% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{UDM_test}
\alias{UDM_test}
\title{Test with unnormalized difference-in-means test statistic}
\usage{
UDM_test(data)
}
\arguments{
\item{data}{Data to be analyzed}
}
\value{
left, right and both sided tests
}
\description{
Test with unnormalized difference-in-means test statistic
}
