% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{weighted_AIPW}
\alias{weighted_AIPW}
\title{Tnis is a function building up the point and variance estimates wight weighting strategy}
\usage{
weighted_AIPW(data, hyperparams, reg_type = "lasso")
}
\arguments{
\item{data}{Same as AIPW function}

\item{hyperparams}{Same as AIPW function}

\item{reg_type}{Same as AIPW function}
}
\value{
An output matrix with point and variance estimate for each arm
}
\description{
Tnis is a function building up the point and variance estimates wight weighting strategy
}
