% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{weighted_IPW}
\alias{weighted_IPW}
\title{Tnis is a function building up the point and variance estimates wight weighting strategy}
\usage{
weighted_IPW(data)
}
\arguments{
\item{data}{Same as AIPW function}
}
\value{
An output matrix with point and variance estimate for each arm
}
\description{
Tnis is a function building up the point and variance estimates wight weighting strategy
}
